organize commands into files, add logic around preventing more than one poll
