+++ commands/disconnect.py
@@ -1,0 +1,6 @@
+# import discord
+
+async def disconnect(client, message):
+	for vc in client.voice_clients:
+		await vc.disconnect()
+		return
