+++ var.py
@@ -1,0 +1,4 @@
+token = 'NDcwMzE5NjkzMjk1NjQ4NzY4.Dj6w2A.IS0A5Gt2TL7yFpW36FQ1vBOe9i4'
+welcome_channel_id = '470289384814542860'
+is_polling = False
+current_poll_owner = None
